removed separate graph database
