+++ CHANGELOG.md
@@ -10,7 +10,8 @@ Write the date in place of the "Unreleased" in the case a new version is release
 - Support clustered Redis for high availability.
 - Add client_secret,redirect_on_success,redirect_on_failure to
   ProxiedOIDCAuthenticator. This is to allow login using Tiled-UI
-- Allow configuration of user_id_claim for OIDCAuthenticator.
+- Allow configuration of user_id_claim for OIDCAuthenticator
+- Adds a graph of links triplestore-light feature to tiled using the catalog database
 
 ### Fixed
