+++ tiled/catalog/core.py
@@ -6,6 +6,7 @@ from .base import Base
 
 # This is list of all valid revisions (from current to oldest).
 ALL_REVISIONS = [
+    "c31f6a1d7e20",
     "9bc9b57294b9",
     "b93c79d197f4",
     "e8956581ecd5",
